+++ test/dummy/app/components/fake_block/component.rb
@@ -1,0 +1,6 @@
+class FakeBlock::Component < ViewComponent::Base
+  def initialize(recording:)
+    @recording = recording
+    @fake_block = recording.recordable
+  end
+end